New Java signing-parent and signing-pkcs11 projects; first PKCS#11 signing test
itext-signing-parent is the Java parent project of the signing projects.
itext-signing-pkcs11 is the Java project focusing on signing via PKCS#11.

The IExternalSignature implementation Pkcs11Signature handles the PKCS#11 related initialization in excess to the actual signing.
README.md explains the use of SoftHSM2 as PKCS#11 backend.
pkcs11.cfg is an example PKCS#11 configuration file.
TestPkcs11Access tests accessibility of the SoftHSM2 backend.
TestSignSimple tests the creation of a simple PDF signature via the SoftHSM2 backend via a Pkcs11Signature instance.