Generalized Pkcs11Signature for keys with certificates with different labels
Pkcs11Signature.Select now expects an extra string parameter certLabel.
If it is not null, it serves as a CKA_LABEL filter for the certificate.
If it is null, a non-null private key CKA_LABEL is used instead.
If that also is null, a non-null CKA_ID is used instead.
If that also is null, no CKA_LABEL filter for the certificate is applied.

TestPkcs11SignSimple is updated accordingly. Furthermore, it is prepared for adding signature timestamp support.