New Java signing-simple project
itext-signing-simple is the Java project focusing on simple signing, e.g. using generic JCA/JCE private keys and X509 certificates.

PrivateKeySignatureContainerBC allows signing a PDF with iText 7 using a JCS/JCE private key and X509 certificate. In contrast to the PrivateKeySignature already included in iText 7 this class also generates the CMS container to be embedded in the PDF. Thus, it in particular is not restricted by the limitations of the CMS container generation code used by iText itself.

TestSignBCSimple tests the PrivateKeySignatureContainerBC, in particular also creating a RSASSA-PSS signature which is not supported by the integrated CMS signature container generation of iText.