Created UtimacoJceSignatureContainer in itext-signing-jce-utimaco
Unfortunately the PrivateKeySignatureContainerBC from the itext-signing-simple project cannot be used to create RSASSA-PSS signatures with the Utimaco Java security provider: The latter only accepts regular SHAxxxWITHRSA signature algorithm names and for RSASSA-PSS expects the AlgorithmParameterSpec parameter to be a PSSParameterSpec. The former, though, uses a ContentSigner instance created by the BouncyCastle JcaContentSignerBuilder, but this builder class transforms the given signature algorithm name forth and back so that the underlying provider is asked for SHAxxxWITHRSAANDMGF1 whether one started with SHAxxxWITHRSA or with SHAxxxWITHRSAANDMGF1.

In contrast to that the UtimacoJceSignatureContainer builds its own ContentSigner and uses exactly the given arguments when requesting a signature from the Utimaco provider.

The new TestSignSimple tests testSignSimpleUtimacoJceSignatureContainerRsaPkcs1 and testSignSimpleUtimacoJceSignatureContainerRsaSsaPss illustrate the usage.